requirements.txt
python-binance==1.0.19
requests==2.32.3
httpx==0.27.2
click==8.1.7
typer==0.12.5
pytest==8.3.3
python-dotenv==1.0.1
